backend/requirements.txt
Django
djangorestframework
ffmpeg-python
numpy
opencv-python
audio-offset-finder
